+++ models.py
@@ -10,6 +10,7 @@ class User(db.Model):
     __tablename__ = 'users'
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(50), unique=True, nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)  # Email wajib untuk login
     password = db.Column(db.String(128), nullable=False)
     role = db.Column(db.String(10), nullable=False, default='user')  # admin / user
     active_until = db.Column(db.DateTime, nullable=True)